+++ test/unit/test_test.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from prysk.run import _findtests, cwd
+from prysk.test import _findtests, cwd
 
 
 def create_directory(root, name):
